PR feedback: use correct page name
